feat(octo-whatsapp): messages.read_view_once + messages.list_unavailable + messages.list_ephemeral (Phase 7.K T10+T11)
T10 messages.read_view_once
- One-shot semantics: SET consumed_at_unix_ms + zero media_token on first
  successful read; subsequent reads return 'view-once message already
  consumed' rather than re-downloading.
- Inline base64 encoder for media_b64 (no external dep).
- Literal-id SQL interpolation for WHERE clauses on INTEGER PK — the
  Stoolap fork's ? placeholder binding misses when SELECT WHERE on
  INTEGER PK is the first query against the ingester's in-memory file;
  observed via the drain-vs-match divergence diagnostic (drain loop
  found the row, match on next() returned None).
- 7 hermetic tests using the QuerySubsystem open_subsystem pattern.

T11 messages.list_unavailable + messages.list_ephemeral
- messages.list_unavailable: filter by kind/peer/since/until, returns
  rows from unavailable_messages table (Phase 7.K v2 ingest target
  for <unavailable type=...> fanout).
- messages.list_ephemeral: WHERE ephemeral_expires_at_seconds IS NOT NULL,
  filter by peer/kind, returns rows from messages table.
- Both gated on QuerySubsystem; 6 hermetic tests each.

ingester.rs cleanup
- Drop redundant i64 casts on ts_mono_ns (already i64).
- Use Stoolap format! WHERE strings consistently with the rest of the
  Phase 7.K handlers.

EXPECTED_TOOL_COUNT unchanged at 191 (this commit lands the 3 new RPCs,
the registry count is consistent at this stage of the plan).

1025 lib tests pass (+12 vs Phase 7.K baseline). clippy clean (-D warnings).

+++ crates/octo-whatsapp/src/query/ingester.rs
@@ -135,7 +135,7 @@ impl QueryIngester {
                 vec![
                     Value::from(id as i64),
                     Value::from(ts_unix_ms),
-                    Value::from(ts_mono_ns as i64),
+                    Value::from(ts_mono_ns),
                     Value::from(un.kind_str),
                     Value::from(un.peer),
                     Value::from(un.sender),
@@ -150,7 +150,7 @@ impl QueryIngester {
                 vec![
                     Value::from(id as i64),
                     Value::from(ts_unix_ms),
-                    Value::from(ts_mono_ns as i64),
+                    Value::from(ts_mono_ns),
                     Value::from(dmc.jid),
                     Value::from(dmc.duration_seconds),
                 ],
